Fixed all except pico issue
All read and writes appear to work as intended barring when pico
attempts to read from an offset of 0 up to a given part of the file. As
of now read function only takes into account the offset to start at and
will read to the end of the file regardless of all other factors.